Get started on a tool to update links in blog posts
Many of my blog posts have relative links to other posts or resources on my
site. But my url scheme has changed over the years and some of the links refer
to old names for pages. These are all equipped with redirects but the presence
of the old links makes crawlers anxious, so this script tries to find links and
decode them and decide if there's an update.